Restructure: pytorch training as process without further subdirectories
